docs(setup): add installation setup instructions
As part of the use of this repository, it is important to get up and
running with it to test.  As such, steps will need to be taken to ensure
that the dependencies required to run the tests are present, as well as
the browsers playwright uses to run these tests through.  This is the
start of documenting requirements and installation instructions for
using the project locally.

Note that this assumes a Linux/MacOS environment.