feat(sitemap): tests against sitemap retrieval
Fetch is an async method, and playwright needs synchronous code to
properlyl be able to write the tests dynamically.  As such, a bash file
was created to retrieve the locations from the sitemap.xml.

To run:

`npm run sitemap https://mysiteIwantSitemapFrom`: Retrieves the sitemap
to sitemap.links.

`npm run a11y https://mysiteIwantSitemapFrom`: Retrieves the sitemap to
a local file and runs the axe tests for it.

This will generate a file named `sitemap.links`, which axe-spec uses
to get the entries from.  Since file IO is blocking, this allows the
axe-sitemap.spec.ts tests to properly be created.

* Adds a bash script to retrieve the sitemap list from a website for use in
  axe testing.
* Adds javascript version of sitemap retrieval should we want to use
  that instead.
* This also ignores sitemap.links through gitignore.