docs(core): update readme with local setup instructions (#8)
* docs(setup): add installation setup instructions

As part of the use of this repository, it is important to get up and
running with it to test.  As such, steps will need to be taken to ensure
that the dependencies required to run the tests are present, as well as
the browsers playwright uses to run these tests through.  This is the
start of documenting requirements and installation instructions for
using the project locally.

Note that this assumes a Linux/MacOS environment.

* docs(sitemaps): add instructions for running sitemap tests

Now that PR #9 is introduced, we need documentation on how to use this
sitemap pulling functionality to test sites on a remote server.  This
adds that documentation.

* docs(requirements): add xmllint as requirement

Since we use xmllint in the bash version of the sitemap pull, we will
include it here.  We could swap to the js version later to drop this
requirement.