Documentation and Visual Regression Testing (#10)
* docs(visual): add images for docs

* feat(visreg): add example visual regression test

* feat(visreg): use sitemap to run visual regression tests

* feat(visreg): add visreg testing

Adds a way to visual regression test a sitemap list of pages.

The first time you run this, it will take snapshots.  The next times you
run it, it will compare to the original.  To create a new base, delete
the folder with the screenshots and run again.

* feat(targets): target desktop chrome and mobile safari

While you can modify to test against multiple browsers, we set defautls
so that the tests don't take as long and to still give the optimal
coverage.

* docs(visreg): add visual regression documentation

---------

Co-authored-by: Michael Vaughn <michael.vaughn@yale.edu>