feat(axe): provide better output of findings
Before, we would output issues found in the axe-sitemap to the stdout,
resulting in an attachment of sorts toward the bottom of a test for that
page.  This caused the developer to have to scroll down to expand the
stdout to see the error.  Now this is done as part of the reporter
findings, making it much easier to see what needs to be done without the
need for stdout.

This also introduces the optional use of an environment variable to
serve as the BASE_URL for any test you would write.  For instance, you
could set:

`TEST_URL=http://localhost:3000 npx playwright test <file>` and it would
use the TEST_URL provided as the base url.  This results in not needing
to specify the whole URL in the page visits of your tests.  Instead, you
can use the relative path and it'll prepend the base URL.

This also allows retrieveSitemap to allow insecure sites in the event
you have a self signed certificate locally, but would still like to test
it.